+++ AverageColorLib/AverageColor_ParallelTransform.h
@@ -2,8 +2,8 @@
 
 #include "RawBitmap.h"
 
-unsigned long AverageColor_ParallelTransform(
+DWORD AverageColor_ParallelTransform(
         RawBitmap::PixelColorVector::iterator begin,
         RawBitmap::PixelColorVector::iterator end,
-        unsigned int width,
-        unsigned int height);
+        UINT width,
+        UINT height);
