+++ AverageColorLib/ColorCalculations.h
@@ -3,10 +3,10 @@
 #include <array>
 #include "RawBitmap.h"
 
-typedef std::array<unsigned long long, sizeof(RawBitmap::PixelColor)> PixelColorSums;
+typedef std::array<ULONGLONG, sizeof(RawBitmap::PixelColor)> PixelColorSums;
 
 PixelColorSums AddPixelColorToSums(PixelColorSums pixelColorSums, const RawBitmap::PixelColor& pixelColor);
-unsigned long CalculateAverage(const PixelColorSums& totals, unsigned int pixelCount);
+DWORD CalculateAverage(const PixelColorSums& totals, UINT pixelCount);
 
 inline PixelColorSums operator+(PixelColorSums lhs, const PixelColorSums& rhs)
 {
